refactor(scene_protocol): ImageAnnotations as first-class builtin
Promote ImageAnnotations from a typedef alias into a fully-defined SDK
builtin, aligned with the shape of Image / DepthImage / PointCloud.

- Define struct PJ::sdk::ImageAnnotations and its companions
  (Point2, ColorRGBA, AnnotationTopology, PointsAnnotation,
  CircleAnnotation, TextAnnotation) directly in
  pj_scene_protocol/builtin/ImageAnnotations.h.
- Delete the legacy pj_scene_protocol/image_annotation.h.
- Move SceneFrame (codec output wrapper) into a dedicated
  pj_scene_protocol/scene_frame.h, now wrapping
  std::vector<sdk::ImageAnnotations>.
- Codec: serializeImageAnnotation -> serializeImageAnnotations,
  signature switched to PJ::sdk::ImageAnnotations.
- Update consumers (codec impl, decoder, tests, docs).

Build green; 69/69 PJ4 tests pass.
