+++ pj_scene_protocol/tests/image_annotation_codec_test.cpp
@@ -7,11 +7,19 @@
 #include <string>
 #include <vector>
 
-#include "pj_scene_protocol/image_annotation.h"
 #include "pj_scene_protocol/scene_decoder.h"  // existing reader, used for round-trips
+#include "pj_scene_protocol/scene_frame.h"
 
 namespace PJ {
 namespace {
+
+using sdk::AnnotationTopology;
+using sdk::CircleAnnotation;
+using sdk::ColorRGBA;
+using sdk::ImageAnnotations;
+using sdk::Point2;
+using sdk::PointsAnnotation;
+using sdk::TextAnnotation;
 
 // -----------------------------------------------------------------------------
 // Hand-rolled Protobuf helpers — same style as the sibling decoder test
@@ -47,11 +55,11 @@ inline void appendLenDelim(std::vector<uint8_t>& out, const std::vector<uint8_t>
 
 }  // namespace pb
 
-// Decode the bytes produced by serializeImageAnnotation back into an
-// ImageAnnotation. Returns the inner annotation; assumes the SceneFrame wraps
-// exactly one ImageAnnotation (the reader's contract).
-ImageAnnotation roundTrip(const ImageAnnotation& input) {
-  auto bytes = serializeImageAnnotation(input);
+// Decode the bytes produced by serializeImageAnnotations back into an
+// sdk::ImageAnnotations. Returns the inner annotation; assumes the SceneFrame wraps
+// exactly one sdk::ImageAnnotations (the reader's contract).
+sdk::ImageAnnotations roundTrip(const sdk::ImageAnnotations& input) {
+  auto bytes = serializeImageAnnotations(input);
   auto decoder = makeSceneDecoder(kSchemaImageAnnotations);
   EXPECT_NE(decoder.get(), nullptr);
   auto result = decoder->decode(bytes.data(), bytes.size());
@@ -76,8 +84,8 @@ ImageAnnotation roundTrip(const ImageAnnotation& input) {
 // -----------------------------------------------------------------------------
 
 TEST(ImageAnnotationCodecTest, EmptyAnnotationProducesEmptyBytes) {
-  ImageAnnotation ia;
-  auto bytes = serializeImageAnnotation(ia);
+  sdk::ImageAnnotations ia;
+  auto bytes = serializeImageAnnotations(ia);
   EXPECT_TRUE(bytes.empty());
 }
 
@@ -89,7 +97,7 @@ TEST(ImageAnnotationCodecTest, GoldenBytes_SinglePointsAnnotation) {
   // Build the canonical input: one PointsAnnotation, kLineLoop, two points,
   // outline color = pure red (255, 0, 0, 255), fill = transparent default,
   // thickness = 2.0.
-  ImageAnnotation ia;
+  sdk::ImageAnnotations ia;
   PointsAnnotation pa;
   pa.topology = AnnotationTopology::kLineLoop;
   pa.points = {{1.0, 2.0}, {3.0, 4.0}};
@@ -156,7 +164,7 @@ TEST(ImageAnnotationCodecTest, GoldenBytes_SinglePointsAnnotation) {
   pb::appendTag(expected, 2, 2);
   pb::appendLenDelim(expected, pa_body);
 
-  auto actual = serializeImageAnnotation(ia);
+  auto actual = serializeImageAnnotations(ia);
   EXPECT_EQ(actual, expected) << "wire format mismatch";
 }
 
@@ -165,7 +173,7 @@ TEST(ImageAnnotationCodecTest, GoldenBytes_SinglePointsAnnotation) {
 // -----------------------------------------------------------------------------
 
 TEST(ImageAnnotationCodecTest, RoundTrip_LineLoopFourPoints) {
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
   PointsAnnotation pa;
   pa.topology = AnnotationTopology::kLineLoop;
   pa.points = {{75.0, 185.0}, {125.0, 185.0}, {125.0, 215.0}, {75.0, 215.0}};
@@ -187,7 +195,7 @@ TEST(ImageAnnotationCodecTest, RoundTrip_AllTopologies) {
   for (auto topology :
        {AnnotationTopology::kPoints, AnnotationTopology::kLineList, AnnotationTopology::kLineStrip,
         AnnotationTopology::kLineLoop}) {
-    ImageAnnotation in;
+    sdk::ImageAnnotations in;
     PointsAnnotation pa;
     pa.topology = topology;
     pa.points = {{0.0, 0.0}, {10.0, 10.0}};
@@ -203,7 +211,7 @@ TEST(ImageAnnotationCodecTest, RoundTrip_AllTopologies) {
 }
 
 TEST(ImageAnnotationCodecTest, RoundTrip_CirclePreservesDiameterRadiusInverse) {
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
   CircleAnnotation ca;
   ca.center = {50.0, 60.0};
   ca.radius = 10.0;  // wire emits diameter = 20; reader halves back to 10.
@@ -223,7 +231,7 @@ TEST(ImageAnnotationCodecTest, RoundTrip_CirclePreservesDiameterRadiusInverse) {
 }
 
 TEST(ImageAnnotationCodecTest, RoundTrip_TextUtf8) {
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
   TextAnnotation ta;
   ta.position = {320.5, 240.25};
   ta.font_size = 14.0;
@@ -241,7 +249,7 @@ TEST(ImageAnnotationCodecTest, RoundTrip_TextUtf8) {
 }
 
 TEST(ImageAnnotationCodecTest, RoundTrip_FullImageAnnotationAllThreeKinds) {
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
 
   // Two points annotations.
   PointsAnnotation pa1;
@@ -292,7 +300,7 @@ TEST(ImageAnnotationCodecTest, EmptyColorsVectorDoesNotInjectDefaultEntry) {
   // A PointsAnnotation with empty `colors` must round-trip to empty `colors`.
   // If the writer emitted a default Color for the empty vector, the reader
   // would push a phantom entry, breaking per-vertex coloring semantics.
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
   PointsAnnotation pa;
   pa.topology = AnnotationTopology::kPoints;
   pa.points = {{1.0, 1.0}, {2.0, 2.0}};
@@ -307,7 +315,7 @@ TEST(ImageAnnotationCodecTest, EmptyColorsVectorDoesNotInjectDefaultEntry) {
 }
 
 TEST(ImageAnnotationCodecTest, RoundTrip_PerVertexColors) {
-  ImageAnnotation in;
+  sdk::ImageAnnotations in;
   PointsAnnotation pa;
   pa.topology = AnnotationTopology::kLineStrip;
   pa.points = {{0.0, 0.0}, {10.0, 10.0}, {20.0, 0.0}};
